# simplify cooperations and defections2