Turn them in to properties and read them from player.history
Makes code simpler