@drvinceknight 2015-10-02
Yeah this simplifies things: I like the look of it. 

Minor question: does this not though (arguably) slow things down a bit as we have to count the C's and D's every time as opposed to just keeping count of them as we go?