@drvinceknight 2015-10-02
(Speed is not a terribly sensitive issue so it doesn't mean that it's a no go: just thinking about things)