@marcharper 2015-10-02
It's O(rounds) vs O(rounds ^ 2) for the strategies that use the properties every round. The impact will depend on what strategies are in use.

We could run `time ./run_axelrod` as a simple benchmark (with and without the patch).