@bootandy 2015-10-03
Yes it would slow things down slightly if the other player accessed the property on their strategy call. (Most current strategies dont).

I realise I have added something that does turn O(rounds) into O(rounds^2) which from a pure computer science point of view is a bad thing. 

However this should be balanced against not having to update 4 more variables after each round. And the fact the number of rounds is often low.

As long as the number of rounds a pair of player plays is low (say ~100) I do not think it will have a negative performance impact. However I would like to see the results of the ./run_axelrod benchmark as I may be wrong.