@meatballs 2015-10-03
@bootandy I think the simplified code outweighs the performance. If lots of strategies start using it, then that view may change but, at the moment, I like the look of what you've done.

Not sure if you know, but we have a chat channel on gitter which you're welcome to join: https://gitter.im/Axelrod-Python/Axelrod