@bootandy 2015-10-03
This branch:
 Starting basic_strategies tournament with 10 round robins of 200 turns per pair.
 Finished all basic_strategies tasks in 1.7s
 Starting strategies tournament with 10 round robins of 200 turns per pair.
 Finished all strategies tasks in 166.5s
 Starting cheating_strategies tournament with 10 round robins of 200 turns per pair.
 Finished all cheating_strategies tasks in 6.5s
  
Old code:
git checkout 0a3a9aa
 
  ./run_axelrod
 Starting basic_strategies tournament with 10 round robins of 200 turns per pair.
 Finished all basic_strategies tasks in 1.7s
 Starting strategies tournament with 10 round robins of 200 turns per pair.
 Finished all strategies tasks in 150.2s
 Starting cheating_strategies tournament with 10 round robins of 200 turns per pair.
 Finished all cheating_strategies tasks in 6.4s

So 150s -> 166s ~ 10% performance decrease. 
I think it is worth it for simpler code and less redundant data.