@drvinceknight 2015-10-03
I agree with Owen re readability over performance, and those timing impacts
are minor. :+1:

On Sat, 3 Oct 2015 12:45 andy boot <notifications@github.com> wrote:

> This branch:
> Starting basic_strategies tournament with 10 round robins of 200 turns per
> pair.
> Finished all basic_strategies tasks in 1.7s
> Starting strategies tournament with 10 round robins of 200 turns per pair.
> Finished all strategies tasks in 166.5s
> Starting cheating_strategies tournament with 10 round robins of 200 turns
> per pair.
> Finished all cheating_strategies tasks in 6.5s
>
> Old code:
> git checkout 0a3a9aa
> <https://github.com/Axelrod-Python/Axelrod/commit/0a3a9aab0786af752e5e25cb11c63da776c7ab1f>
>
> ./run_axelrod
> Starting basic_strategies tournament with 10 round robins of 200 turns per
> pair.
> Finished all basic_strategies tasks in 1.7s
> Starting strategies tournament with 10 round robins of 200 turns per pair.
> Finished all strategies tasks in 150.2s
> Starting cheating_strategies tournament with 10 round robins of 200 turns
> per pair.
> Finished all cheating_strategies tasks in 6.4s
>
> So 150s -> 166s ~ 10% performance decrease.
> I think it is worth it for simpler code and less redundant data.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/Axelrod-Python/Axelrod/pull/334#issuecomment-145242235>
> .
>