@marcharper 2015-10-03
I'm not sure I agree that this is more simple. It's shorter, of course, but only because it's hiding the calculations in `history.count`. It's computationally more complex, as we all agree.

As for readability, we could alternatively have a History object that keeps the history and updates a `collections.Counter` as the history grows. This object could be instantiated by an existing history (for easy restoration in eg. MindReader, or just copied), and the two attributes `.cooperations` and `.defections` could just call `counter['C']` and `counter['D']`, keeping the property interface for `Player` (or moving it to `Player.history.cooperations`). This also preserves the performance advantage, and allows for future changes/additions to the history without messing with the `Player` class.