@drvinceknight 2015-10-04
> Edit: Making the history a class isn't easy at this point -- there are a lot of places in the code that rely on some aspect of the current implementation (e.g. shared history for MetaPlayer). I left my progress here if anyone else wants to give it a go.

I still think that this is more readable (as you say it hides the complexity in `count` which is fairly verbose).  Perhaps and approach could be to merge this as is (I think we're all relatively comfortable with the increased run time?) and perhaps at a later point (once @meatballs has finished his big refactor) we could bring in the History class?