@drvinceknight 2015-10-05
What do you think of the above suggestion @marcharper ?