@drvinceknight 2015-10-05
On reflection perhaps this can be shelved for now as it would be nicer perhaps to just include whatever the final solution is.

Really appreciate the contribution @bootandy :+1: :)